week3 - [refactor] move LottoConstants.kt to .constants package
