+++ khjin-tdd/src/main/kotlin/com/khjin/tdd_practice/week2/lotto/Lotto.kt
@@ -1,6 +1,7 @@
 package com.khjin.tdd_practice.week2.lotto
 
 import com.khjin.tdd_practice.week2.WinningNumber
+import com.khjin.tdd_practice.week2.lotto.constants.LottoConstants
 
 class Lotto {
