+++ khjin-tdd/src/main/kotlin/com/khjin/tdd_practice/week2/lotto/LottoValidator.kt
@@ -1,5 +1,6 @@
 package com.khjin.tdd_practice.week2.lotto
 
+import com.khjin.tdd_practice.week2.lotto.constants.LottoConstants
 import com.khjin.tdd_practice.week2.lotto.exception.InvalidWinnerInputException
 
 class LottoValidator {
